+++ src/selection_indicator/style/line.rs
@@ -5,7 +5,13 @@ use embedded_graphics::{
     Drawable,
 };
 
-use crate::selection_indicator::{style::IndicatorStyle, Insets};
+use crate::{
+    interaction::InputState,
+    selection_indicator::{
+        style::{interpolate, IndicatorStyle},
+        Insets,
+    },
+};
 
 #[derive(Clone, Copy)]
 pub struct Line;
@@ -33,13 +39,19 @@ impl IndicatorStyle for Line {
     fn draw<D>(
         &self,
         state: &Self::State,
-        fill_width: u32,
+        input_state: InputState,
         display: &mut D,
     ) -> Result<u32, D::Error>
     where
         D: DrawTarget<Color = BinaryColor>,
     {
         let display_area = display.bounding_box();
+
+        let fill_width = if let InputState::InProgress(progress) = input_state {
+            interpolate(progress as u32, 0, 255, 0, display_area.size.width)
+        } else {
+            0
+        };
 
         self.shape(state, display_area, fill_width)
             .into_styled(PrimitiveStyle::with_fill(BinaryColor::On))