+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Unreleased (??)
  - `Menu::add_items` now accepts owning collections (e.g. `Vec`)
  - Replaced `StyledDrawable` wtih `MenuItem::draw_styled` and `MenuItemCollection::draw_styled`
  - Empty menu titles are no longer displayed
+ - Reworked input handling. `InteractionController` has been replaced by `InputAdapter`
 
 0.3.1 (2023-08-06)
 ==================